backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Maxim <mlyulenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
drf-yasg = "^1.21.7"
pillow = "^10.0.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.2.0"
celery = "^5.3.4"
django-countries = "^7.5.1"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
audio-metadata = "^0.11.1"
django-polymorphic = "^3.1.0"
django-filter = "^23.3"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
